Extract StoragePersister interface
