+++ app/core/src/main/java/com/fsck/k9/KoinModule.kt
@@ -10,6 +10,7 @@ import com.fsck.k9.mail.ssl.TrustedSocketFactory
 import com.fsck.k9.mailstore.LocalStoreProvider
 import com.fsck.k9.mailstore.StorageManager
 import com.fsck.k9.power.TracingPowerManager
+import com.fsck.k9.preferences.K9StoragePersister
 import com.fsck.k9.preferences.StoragePersister
 import org.koin.dsl.module.applicationContext
 
@@ -17,7 +18,7 @@ val mainModule = applicationContext {
     bean { Preferences.getPreferences(get()) }
     bean { get<Context>().resources }
     bean { StorageManager.getInstance(get()) }
-    bean { StoragePersister(get()) }
+    bean { K9StoragePersister(get()) as StoragePersister }
     bean { LocalStoreProvider() }
     bean { TracingPowerManager.getPowerManager(get()) as PowerManager }
     bean { Contacts.getInstance(get()) }
