+++ app/core/src/main/java/com/fsck/k9/preferences/StoragePersister.kt
@@ -1,0 +1,11 @@
+package com.fsck.k9.preferences
+
+import android.support.annotation.CheckResult
+
+
+interface StoragePersister {
+    @CheckResult
+    fun loadValues(): Map<String, String>
+
+    fun createStorageEditor(storage: Storage): StorageEditor
+}
