+++ app/core/src/test/java/com/fsck/k9/preferences/StoragePersisterTest.kt
@@ -3,8 +3,8 @@ package com.fsck.k9.preferences
 
 import android.content.Context
 import com.fsck.k9.K9RobolectricTest
-import com.fsck.k9.preferences.StoragePersister.StoragePersistOperationCallback
-import com.fsck.k9.preferences.StoragePersister.StoragePersistOperations
+import com.fsck.k9.preferences.K9StoragePersister.StoragePersistOperationCallback
+import com.fsck.k9.preferences.K9StoragePersister.StoragePersistOperations
 import com.nhaarman.mockito_kotlin.*
 import org.junit.Assert.*
 import org.junit.Test
@@ -13,7 +13,7 @@ import org.robolectric.RuntimeEnvironment
 
 class StoragePersisterTest : K9RobolectricTest() {
     private var context: Context = RuntimeEnvironment.application
-    private var storagePersister = StoragePersister(context)
+    private var storagePersister = K9StoragePersister(context)
 
     @Test
     fun doInTransaction_order() {
